Sign in and sign up Complete
Sign in creates a graph named qurantracker. Also saves the info to the .env for future use.
Sign up successfully signs up with the given name and token. Shows any error to the user. Does not save the info
to the .env file. User will have to sign in with their new account.
Asks to open up the users profile when signig up is successful.